+++ packages/salesforcedx-vscode-org/src/commands/orgList.ts
@@ -25,8 +25,7 @@ import {
 import { channelService } from '../channels';
 import { nls } from '../messages';
 import { PromptConfirmGatherer } from '../parameterGatherers/promptConfirmGatherer';
-import { getAuthFieldsFor } from '../util/orgUtil';
-import { getConnectionStatusFromError, shouldRemoveOrg } from '../util/orgUtils';
+import { getAuthFieldsFor, getConnectionStatusFromError, shouldRemoveOrg } from '../util/orgUtil';
 import { OrgListCleanExecutor } from './orgListCleanExecutor';
 
 /** Check actual connection status by testing the connection */
