+++ packages/salesforcedx-vscode-org/src/util/orgUtil.ts
@@ -165,3 +165,30 @@ export const setTargetOrgOrAlias = async (usernameOrAlias: string): Promise<void
     process.chdir(originalDirectory);
   }
 };
+
+/** Get connection status from error */
+export const getConnectionStatusFromError = (err: any, username?: string): string => {
+  const message = err instanceof Error ? err.message : String(err);
+  const lowerMsg = message.toLowerCase();
+
+  if (lowerMsg.includes('maintenance')) return 'Down (Maintenance)';
+  if (lowerMsg.includes('<html>') || lowerMsg.includes('<!doctype html>')) return 'Bad Response';
+  if (
+    ['expired access/refresh token', 'invalid_session_id', 'bad_oauth_token', 'refreshtokenautherror'].some(token =>
+      lowerMsg.includes(token)
+    )
+  ) {
+    return 'Unable to refresh session: expired access/refresh token';
+  }
+  if (shouldRemoveOrg(err)) {
+    return username ? `Invalid org: ${username}` : 'Invalid org';
+  }
+
+  return message;
+};
+
+/** Check if org should be removed based on error */
+export const shouldRemoveOrg = (err: any): boolean => {
+  const lowerMsg = (err instanceof Error ? err.message : String(err)).toLowerCase();
+  return ['invalid_login', 'no such org', 'namedorgnotfound', 'noauthinfofound'].some(msg => lowerMsg.includes(msg));
+};
